Refactor, Additions - Fix problem with saving and reading time - Refactor Navigation Drawer - Add deleting selected items - Refactor note fields type (id = int, time = long)

+++ app/src/main/java/com/stslex93/notes/ui/search/SearchFragment.kt
@@ -117,12 +117,7 @@ class SearchFragment : BaseFragment() {
         }, {})
 
     private fun View.navigation(id: String, edit: Boolean) {
-        val direction =
-            SearchFragmentDirections.actionNavSearchToNavEdit(
-                id,
-                edit,
-                id
-            )
+        val direction = SearchFragmentDirections.actionNavSearchToNavEdit(id.toInt(), edit, id)
         if (this is MaterialCardView) this.isTransitionGroup = true
         val extras = FragmentNavigatorExtras(this to this.transitionName)
         findNavController().navigate(direction, extras)
